+++ HaruhiChokuretsuEditor.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HaruhiChokuretsuEditor", "HaruhiChokuretsuEditor\HaruhiChokuretsuEditor.csproj", "{BFC57E2D-4371-4A50-AC60-307A9F19C4DD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HaruhiChokuretsuTests", "HaruhiChokuretsuTests\HaruhiChokuretsuTests.csproj", "{B6B64437-5C36-482A-85D7-419302E0CF23}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HaruhiChokuretsuTests", "HaruhiChokuretsuTests\HaruhiChokuretsuTests.csproj", "{B6B64437-5C36-482A-85D7-419302E0CF23}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HaruhiChokuretsuLib", "HaruhiChokuretsuLib\HaruhiChokuretsuLib.csproj", "{5ADB6BD3-B3D4-40FA-AA6E-66CD279C271B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{B6B64437-5C36-482A-85D7-419302E0CF23}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B6B64437-5C36-482A-85D7-419302E0CF23}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B6B64437-5C36-482A-85D7-419302E0CF23}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5ADB6BD3-B3D4-40FA-AA6E-66CD279C271B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5ADB6BD3-B3D4-40FA-AA6E-66CD279C271B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5ADB6BD3-B3D4-40FA-AA6E-66CD279C271B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5ADB6BD3-B3D4-40FA-AA6E-66CD279C271B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
