+++ HaruhiChokuretsuLib/ArchiveFile.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HaruhiChokuretsuEditor
+namespace HaruhiChokuretsuLib
 {
     public class ArchiveFile<T>
         where T : FileInArchive, new()
