+++ HaruhiChokuretsuLib/GraphicsFile.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HaruhiChokuretsuEditor
+namespace HaruhiChokuretsuLib
 {
     public class GraphicsFile : FileInArchive
     {
