+++ HaruhiChokuretsuTests/HaruhiChokuretsuTests.csproj
@@ -14,13 +14,13 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="inputs\**">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
+    <ProjectReference Include="..\HaruhiChokuretsuLib\HaruhiChokuretsuLib.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\HaruhiChokuretsuEditor\HaruhiChokuretsuEditor.csproj" />
+    <None Update="inputs\**">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
